requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
coloredlogs==15.0.1
contourpy==1.0.5
cycler==0.11.0
flatbuffers==22.9.24
fonttools==4.37.4
humanfriendly==10.0
idna==3.4
invisible-watermark==0.1.5
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.1
mpmath==1.2.1
numpy==1.23.3
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sympy==1.11.1
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
